// src/library/models/src/lib/src/spells/phb/spell-names.ts
export enum Level3SpellNames {
  ANIMATE_DEAD = 'Animate Dead',
  AURA_OF_VITALITY = 'Aura of Vitality',
  BEACON_OF_HOPE = 'Beacon of Hope',
  BESTOW_CURSE = 'Bestow Curse',
  BLINDING_SMITE = 'Blinding Smite',
  BLINK = 'Blink',
  CALL_LIGHTNING = 'Call Lightning',
  CATNAP = 'Catnap',
  CLAIRVOYANCE = 'Clairvoyance',
  CONJURE_ANIMALS = 'Conjure Animals',
  CONJURE_BARRAGE = 'Conjure Barrage',
  COUNTERSPELL = 'Counterspell',
  CREATE_FOOD_AND_WATER = 'Create Food and Water',
  CRUSADERS_MANTLE = 'Crusader\'s Mantle',
  DAYLIGHT = 'Daylight',
  DISPEL_MAGIC = 'Dispel Magic',
  ELEMENTAL_WEAPON = 'Elemental Weapon',
  ENEMIES_ABOUND = 'Enemies abound',
  ERUPTING_EARTH = 'Erupting Earth',
  FEAR = 'Fear',
  FEIGN_DEATH = 'Feign Death',
  FIREBALL = 'Fireball',
  FLAME_ARROWS = 'Flame Arrows',
  FLY = 'Fly',
  GASEOUS_FORM = 'Gaseous Form',
  GLYPH_OF_WARDING = 'Glyph of Warding',
  HASTE = 'Haste',
  HUNGER_OF_HADAR = 'Hunger of Hadar',
  HYPNOTIC_PATTERN = 'Hypnotic Pattern',
  LEOMUNDS_TINY_HUT = 'Leomund\'s Tiny Hut',
  LIFE_TRANSFERENCE = 'Life Transference',
  LIGHTNING_ARROW = 'Lightning Arrow',
  LIGHTNING_BOLT = 'Lightning Bolt',
  MAGIC_CIRCLE = 'Magic Circle',
  MAJOR_IMAGE = 'Major Image',
  MASS_HEALING_WORD = 'Mass Healing Word',
  MELD_INTO_STONE = 'Meld into Stone',
  MELFS_MINUTE_METEORS = 'Melf\'s Minute Meteors',
  NONDETECTION = 'Nondetection',
  PHANTOM_STEED = 'Phantom Steed',
  PLANT_GROWTH = 'Plant Growth',
  PROTECTION_FROM_ENERGY = 'Protection from Energy',
  REMOVE_CURSE = 'Remove Curse',
  REVIVIFY = 'Revivify',
  SENDING = 'Sending',
  SLEET_STORM = 'Sleet Storm',
  SLOW = 'Slow',
  SPEAK_WITH_DEAD = 'Speak with Dead',
  SPEAK_WITH_PLANTS = 'Speak with Plants',
  SPIRIT_GUARDIANS = 'Spirit Guardians',
  STINKING_CLOUD = 'Stinking Cloud',
  SUMMON_LESSER_DEMONS = 'Summon Lesser Demons',
  THUNDER_STEP = 'Thunder Step',
  TIDAL_WAVE = 'Tidal Wave',
  TINY_SERVANT = 'Tiny Servant',
  TONGUES = 'Tongues',
  VAMPIRIC_TOUCH = 'Vampiric Touch',
  WALL_OF_SAND = 'Wall of Sand',
  WALL_OF_WATER = 'Wall of Water',
  WATER_BREATHING = 'Water Breathing',
  WATER_WALK = 'Water Walk',
  WIND_WALL = 'Wind Wall'
}